Make outages overlay work
