+++ components/Outage/Outage.js
@@ -5,17 +5,17 @@ import { formatRegion } from "../../utils";
 
 const Outage = ({ outage: { region, minutes } }) => {
   return (
-    <div className="flex justify-between p-4">
-      <div className="w-1/2 flex items-center space-x-3">
-        <StatusIcon status="identified" />
-        <p>
-          Down from
-          <span className="font-semibold"> {formatRegion(region)}</span>
-        </p>
+    <div className="bg-white shadow-sm rounded divide-y divide-gray-200 my-1">
+      <div className="flex justify-between p-4">
+        <div className="w-full flex items-center space-x-3">
+          <StatusIcon status="identified" />
+          <p>
+            Down from
+            <span className="font-semibold"> {formatRegion(region)} </span>
+            for {minutes} {minutes === 1 ? "minute" : "minutes"}
+          </p>
+        </div>
       </div>
-      <p className="w-1/2 text-gray-700">
-        down for {minutes} {minutes === 1 ? "minute" : "minutes"}
-      </p>
     </div>
   );
 };
