+++ components/Outages/Outages.test.js
@@ -1,0 +1,7 @@
+import { render, screen } from "@testing-library/react";
+
+import Outages from "./Outages";
+
+describe("Outages", () => {
+  test("renders without errors", () => {});
+});
